services/promotion/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyJWT==2.8.0
django-cors-headers==4.3.1
